+++ pkg/innerring/innerring.go
@@ -12,6 +12,7 @@ import (
 	"github.com/nspcc-dev/neo-go/pkg/core/transaction"
 	"github.com/nspcc-dev/neo-go/pkg/crypto/keys"
 	"github.com/nspcc-dev/neo-go/pkg/encoding/fixedn"
+	"github.com/nspcc-dev/neofs-node/misc"
 	"github.com/nspcc-dev/neofs-node/pkg/innerring/config"
 	"github.com/nspcc-dev/neofs-node/pkg/innerring/internal/blockchain"
 	"github.com/nspcc-dev/neofs-node/pkg/innerring/processors/alphabet"
@@ -965,8 +966,8 @@ func New(ctx context.Context, log *logger.Logger, cfg *viper.Viper, errChan chan
 		queueSize: cfg.GetUint32("workers.subnet"),
 	})
 
-	if cfg.GetString("metrics.address") != "" {
-		m := metrics.NewInnerRingMetrics()
+	if cfg.GetString("prometheus.address") != "" {
+		m := metrics.NewInnerRingMetrics(misc.Version)
 		server.metrics = &m
 	}
 
@@ -1009,12 +1010,9 @@ func createListener(ctx context.Context, cli *client.Client, p *chainParams) (ev
 }
 
 func createClient(ctx context.Context, p *chainParams, errChan chan<- error) (*client.Client, error) {
-	// config name left unchanged for compatibility, may be its better to rename it to "endpoints" or "clients"
-	var endpoints []client.Endpoint
+	endpoints := p.cfg.GetStringSlice(p.name + ".endpoints")
 
-	// defaultPriority is a default endpoint priority
-	const defaultPriority = 1
-
+	// deprecated endpoints with priorities
 	section := p.name + ".endpoint.client"
 	for i := 0; ; i++ {
 		addr := p.cfg.GetString(fmt.Sprintf("%s.%d.%s", section, i, "address"))
@@ -1022,15 +1020,7 @@ func createClient(ctx context.Context, p *chainParams, errChan chan<- error) (*c
 			break
 		}
 
-		priority := p.cfg.GetInt(section + ".priority")
-		if priority <= 0 {
-			priority = defaultPriority
-		}
-
-		endpoints = append(endpoints, client.Endpoint{
-			Address:  addr,
-			Priority: priority,
-		})
+		endpoints = append(endpoints, addr)
 	}
 
 	if len(endpoints) == 0 {
@@ -1043,11 +1033,10 @@ func createClient(ctx context.Context, p *chainParams, errChan chan<- error) (*c
 		client.WithLogger(p.log),
 		client.WithDialTimeout(p.cfg.GetDuration(p.name+".dial_timeout")),
 		client.WithSigner(p.sgn),
-		client.WithEndpoints(endpoints...),
+		client.WithEndpoints(endpoints),
 		client.WithConnLostCallback(func() {
 			errChan <- fmt.Errorf("%s chain connection has been lost", p.name)
 		}),
-		client.WithSwitchInterval(p.cfg.GetDuration(p.name+".switch_interval")),
 	)
 }