+++ pkg/local_object_storage/writecache/storage.go
@@ -5,8 +5,8 @@ import (
 	"fmt"
 	"os"
 
-	lru "github.com/hashicorp/golang-lru"
-	"github.com/hashicorp/golang-lru/simplelru"
+	lru "github.com/hashicorp/golang-lru/v2"
+	"github.com/hashicorp/golang-lru/v2/simplelru"
 	"github.com/nspcc-dev/neofs-node/pkg/local_object_storage/blobstor/common"
 	"github.com/nspcc-dev/neofs-node/pkg/local_object_storage/blobstor/fstree"
 	storagelog "github.com/nspcc-dev/neofs-node/pkg/local_object_storage/internal/log"
@@ -28,7 +28,7 @@ type store struct {
 	// frequently read ones.
 	// MUST NOT be used inside bolt db transaction because it's eviction handler
 	// removes untracked items from the database.
-	flushed simplelru.LRUCache
+	flushed simplelru.LRUCache[string, bool]
 	db      *bbolt.DB
 
 	dbKeysToRemove []string
@@ -74,7 +74,7 @@ func (c *cache) openStore(readOnly bool) error {
 	// Write-cache can be opened multiple times during `SetMode`.
 	// flushed map must not be re-created in this case.
 	if c.flushed == nil {
-		c.flushed, _ = lru.NewWithEvict(c.maxFlushedMarksCount, c.removeFlushed)
+		c.flushed, _ = lru.NewWithEvict[string, bool](c.maxFlushedMarksCount, c.removeFlushed)
 	}
 	return nil
 }
@@ -83,12 +83,11 @@ func (c *cache) openStore(readOnly bool) error {
 // To minimize interference with the client operations, the actual removal
 // is done in batches.
 // It is not thread-safe and is used only as an evict callback to LRU cache.
-func (c *cache) removeFlushed(key, value any) {
-	fromDatabase := value.(bool)
+func (c *cache) removeFlushed(addr string, fromDatabase bool) {
 	if fromDatabase {
-		c.dbKeysToRemove = append(c.dbKeysToRemove, key.(string))
+		c.dbKeysToRemove = append(c.dbKeysToRemove, addr)
 	} else {
-		c.fsKeysToRemove = append(c.fsKeysToRemove, key.(string))
+		c.fsKeysToRemove = append(c.fsKeysToRemove, addr)
 	}
 
 	if len(c.dbKeysToRemove)+len(c.fsKeysToRemove) >= c.maxRemoveBatchSize {