+++ pkg/morph/client/client.go
@@ -8,7 +8,7 @@ import (
 	"sync"
 	"time"
 
-	lru "github.com/hashicorp/golang-lru"
+	lru "github.com/hashicorp/golang-lru/v2"
 	"github.com/nspcc-dev/neo-go/pkg/core/native/noderoles"
 	"github.com/nspcc-dev/neo-go/pkg/core/transaction"
 	"github.com/nspcc-dev/neo-go/pkg/crypto/keys"
@@ -24,7 +24,6 @@ import (
 	"github.com/nspcc-dev/neo-go/pkg/vm/vmstate"
 	"github.com/nspcc-dev/neo-go/pkg/wallet"
 	"github.com/nspcc-dev/neofs-node/pkg/util/logger"
-	"go.uber.org/atomic"
 	"go.uber.org/zap"
 )
 
@@ -59,7 +58,7 @@ type Client struct {
 
 	cfg cfg
 
-	endpoints endpoints
+	endpoints []string
 
 	// switchLock protects endpoints, inactive, and subscription-related fields.
 	// It is taken exclusively during endpoint switch and locked in shared mode
@@ -73,11 +72,6 @@ type Client struct {
 	// establish connection to any of the
 	// provided RPC endpoints
 	inactive bool
-
-	// indicates that Client has already started
-	// goroutine that tries to switch to the higher
-	// priority RPC node
-	switchIsActive atomic.Bool
 }
 
 type cache struct {
@@ -85,7 +79,7 @@ type cache struct {
 
 	nnsHash   *util.Uint160
 	gKey      *keys.PublicKey
-	txHeights *lru.Cache
+	txHeights *lru.Cache[util.Uint256, uint32]
 }
 
 func (c cache) nns() *util.Uint160 {